Delete movie only by admin or owner
Movies can be deleted only by it's owners or admins. Need to implement exception handling for better code responding

+++ ProjektNTP.Application/Services/MovieService.cs
@@ -1,4 +1,7 @@
-﻿using AutoMapper;
+﻿using System.Security.Claims;
+using AutoMapper;
+using Microsoft.AspNetCore.Authorization;
+using ProjektNTP.Application.Authorization.Policies;
 using ProjektNTP.Application.Movie.Dtos;
 using ProjektNTP.Domain.Abstractions;
 
@@ -8,16 +11,20 @@ public class MovieService : IMovieService
 {
     private readonly IMovieRepository _repository;
     private readonly IMapper _mapper;
-
-    public MovieService(IMovieRepository repository, IMapper mapper)
+    private readonly IAuthorizationService _authorizationService;
+    private readonly IUserContextService _userContextService;
+    public MovieService(IMovieRepository repository, IMapper mapper, IAuthorizationService authorizationService, IUserContextService userContextService)
     {
         _repository = repository;
         _mapper = mapper;
+        _authorizationService = authorizationService;
+        _userContextService = userContextService;
     }
 
-    public async Task<Guid> CreateMovie(CreateMovieDto movie)
+    public async Task<Guid> CreateMovie(CreateMovieDto movie, Guid id)
     {
         var movieFromDto = _mapper.Map<Domain.Entities.Movie>(movie);
+        movieFromDto.AddedById = id;
         var movieToAdd = await _repository.CreateMovie(movieFromDto);
         return await Task.FromResult(movieToAdd);
     }
@@ -34,16 +41,25 @@ public class MovieService : IMovieService
         return await Task.FromResult(movieDto);
     }
 
-    public async Task<bool> UpdateMovieById(Guid id, CreateMovieDto movie)
+    public async Task<bool> UpdateMovieById(Guid movieId, CreateMovieDto movieDto)
     {
-        var userToUpdate = _mapper.Map<Domain.Entities.Movie>(movie);
-        var updatedUser = await _repository.UpdateMovieById(id, userToUpdate);
-        return await Task.FromResult(updatedUser);
+        var newMovie = _mapper.Map<Domain.Entities.Movie>(movieDto);
+        var oldMovie = await _repository.GetMovieById(movieId);
+        var updatedMovie = await _repository.UpdateMovieById(movieId, newMovie);
+        return await Task.FromResult(updatedMovie);
     }
 
     public async Task<bool> DeleteMovie(Guid id)
     {
-        var isUserDeleted = await _repository.DeleteMovie(id);
-        return await Task.FromResult(isUserDeleted);
+        var user = _userContextService.User;
+        var userRole = user.Claims.FirstOrDefault(c => c.Type == ClaimTypes.Role).Value;
+        var movie = await _repository.GetMovieById(id);
+        if (movie is null) return await Task.FromResult(false);
+
+        var authorizationResult = _authorizationService.AuthorizeAsync(user, movie, new IsOwnerOrAdminRequirement(IsOwnerOrAdminResourceOperation.Delete)).Result;
+        if (!authorizationResult.Succeeded) throw new UnauthorizedAccessException();
+        
+        await _repository.DeleteMovie(id);
+        return await Task.FromResult(true);
     }
 }
